+++ docs/fr_FR/changelog.md
@@ -1,3 +1,9 @@
+# 23/10/2019
+
+- Mise à jour de la doc
+- Supression forcé des " lors d'un calcul sur une commande
+- Correction de bug de calcul si le résultat à pour valeur"Faux"
+
 # 10/09/2019
 
 - Correction de l'arrondit forcé sur 2 chiffres sur les commandes numérique